pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Polymarket AI trading bot using Claude Opus 4.6"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.34",
    "anthropic>=0.50",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "web3>=7.0",
]
